job_automation/requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==0.8.11
email-validator==2.2.0
